use H2 Database

+++ src/main/java/com/ghanem/RestfulWebServicesApplication/socialmedia/dao/UserDaoService.java
@@ -11,13 +11,14 @@ import java.util.function.Predicate;
 @Component
 public class UserDaoService {
     private  static List<User> users = new ArrayList<User>();
-    private static int usersCount = 0;
+    private static Long usersCount = 0L;
 
     static {
-        users.add(new User(++usersCount,"Adam",LocalDate.now().minusYears(30)));
+        users.add(new User(++usersCount,"Adam", LocalDate.now().minusYears(30)));
         users.add(new User(++usersCount,"Eve",LocalDate.now().minusYears(25)));
         users.add(new User(++usersCount,"Jim",LocalDate.now().minusYears(20)));
     }
+
     public List<User> findAll() {
         return users;
     }
